Moved functions to different packages as suggested by Ben
